Add game funcs
Added some styling to the document and removed placeholders from html.
Added funcs that create players, render the board on the screen, and
update the board state internally every time a box is clicked. Added
minor checks to make sure the content being put into the internal array
is valid