Add end game and check tie funcs
Added logic to check for a tie if all spaces are filled and no winner
has been announced. Added end game func that resets all the internal
boards and variables once the game is complete. Adjusted original
internal game board to be blank in order to fix erroneous tie bugs