Add ui interactivity display and funcs
Added html elements to the header, footer, and sidebar for extra styling
and player interaction. Added css styling in order to make the page look
better. Added funcs that start the game, get player names, and reset the
game according to what button was pressed or if the game has concluded